Move KeyCapturingButton
